refactor: fusionne ui/common/utils et ui/common/hooks
Candidat architecture #2 : ui/utils/ et ui/common/utils/ coexistaient
sans regle de repartition claire (date-utils.ts existait dans les
deux, avec des fonctions differentes et zero chevauchement de nom -
8 fichiers devaient importer les deux a la fois). Meme constat pour
ui/app/hooks/ et ui/common/hooks/ (use-disclosure importe aussi bien
depuis ui/app/** que ui/components/**, aucune scope reelle).

- date-utils.ts fusionne dans ui/utils/ (source unique)
- debounce.ts deplace vers ui/utils/
- use-disclosure, use-user-history-update, use-user-permissions-actions
  deplaces vers ui/app/hooks/
- useLocalStorage delegue desormais a localStorageGet/Set (ui/utils/local-storage.ts)
  au lieu d'appeler localStorage directement : recupere le try/catch
  deja present cote util (JSON corrompu ou quota depasse ne fait plus
  planter le composant), source unique pour "comment on parle au storage".

ui/common/ ne garde que ce qui n'etait pas ambigu : contants.ts,
dayjs.ts, validation/. Les 4 emplacements de hooks scopes par route
(ui/app/beta/_hooks, recherche/_hooks) restent inchanges - convention
Next.js legitime (prefixe _), pas le meme probleme.
