+++ ui/app/(candidat)/emploi/[type]/[id]/[intitule-offre]/CandidaterButton.tsx
@@ -6,7 +6,7 @@ import { Box } from "@mui/material"
 import { useState } from "react"
 import type { ILbaItemLbaCompanyJson, ILbaItemLbaJobJson, ILbaItemPartnerJobJson } from "shared"
 import { isOfferActive } from "@/app/(candidat)/(recherche)/recherche/_components/is-offer-active"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { useSubmitCandidature } from "@/components/ItemDetail/CandidatureLba/services/submit-candidature"
 import ItemDetailApplicationsStatus from "@/components/ItemDetail/ItemDetailServices/ItemDetailApplicationStatus"
 import { notifyJobPostulerV3 } from "@/utils/api"
