+++ ui/app/(espace-pro)/espace-pro/(connected)/_components/CfaHome.tsx
@@ -12,12 +12,12 @@ import type { IRecruiter, IRecruiterJson } from "shared"
 import { Breadcrumb } from "@/app/_components/Breadcrumb"
 import TableWithPagination from "@/app/(espace-pro)/_components/TableWithPagination"
 import { useConnectedSessionClient } from "@/app/(espace-pro)/espace-pro/contexts/userContext"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { useToast } from "@/app/hooks/useToast"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
-import { sortReactTableDate, sortReactTableString } from "@/common/utils/date-utils"
 import { DsfrIcon } from "@/components/DsfrIcon"
 import { AnimationContainer, LoadingEmptySpace } from "@/components/espace_pro"
 import { getEntreprisesManagedByCfa } from "@/utils/api"
+import { sortReactTableDate, sortReactTableString } from "@/utils/date-utils"
 import { PAGES } from "@/utils/routes.utils"
 import { useSearchParamsRecord } from "@/utils/use-search-params-record"
 import { CfaHomeEntrepriseMenu } from "./CfaHomeEntrepriseMenu"
