+++ ui/app/(espace-pro)/espace-pro/(connected)/_components/CompteRenderer.tsx
@@ -10,9 +10,9 @@ import type { CFA, ENTREPRISE } from "shared/constants/recruteur"
 import * as Yup from "yup"
 import CustomInput from "@/app/_components/CustomInput"
 import { useConnectedSessionClient } from "@/app/(espace-pro)/espace-pro/contexts/userContext"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { useToast } from "@/app/hooks/useToast"
 import { AUTHTYPE } from "@/common/contants"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
 import { LoadingEmptySpace } from "@/components/espace_pro"
 import { ArrowRightLine } from "@/theme/components/icons"
 import { getUser, updateUserWithAccountFields } from "@/utils/api"
