+++ ui/app/(espace-pro)/espace-pro/(connected)/_components/DetailEntreprise.tsx
@@ -16,9 +16,9 @@ import Badge from "@/app/(espace-pro)/_components/Badge"
 import { FieldWithValue } from "@/app/(espace-pro)/_components/FieldWithValue"
 import { OpcoSelect } from "@/app/(espace-pro)/_components/OpcoSelect"
 import { useConnectedSessionClient } from "@/app/(espace-pro)/espace-pro/contexts/userContext"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
+import { useUserPermissionsActions } from "@/app/hooks/use-user-permissions-actions"
 import { useToast } from "@/app/hooks/useToast"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
-import { useUserPermissionsActions } from "@/common/hooks/use-user-permissions-actions"
 import { AnimationContainer, ConfirmationDesactivationUtilisateur, ConfirmationModificationOpco, UserValidationHistory } from "@/components/espace_pro"
 import { ArrowRightLine } from "@/theme/components/icons"
 import { updateEntrepriseAdmin, updateEntrepriseCFA } from "@/utils/api"
