+++ ui/app/(espace-pro)/espace-pro/(connected)/_components/EntreprisesGereesParCfa.tsx
@@ -7,9 +7,9 @@ import dayjs from "dayjs"
 import { useState } from "react"
 import type { IRecruiterJson } from "shared"
 import TableWithPagination from "@/app/(espace-pro)/_components/TableWithPagination"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
-import { sortReactTableDate, sortReactTableString } from "@/common/utils/date-utils"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { getEntreprisesManagedByCfa } from "@/utils/api"
+import { sortReactTableDate, sortReactTableString } from "@/utils/date-utils"
 import { ConfirmationSuppressionEntreprise } from "./ConfirmationSuppressionEntreprise"
 import { EntreprisesGereesParCfaMenu } from "./EntreprisesGereesParCfaMenu"
