+++ ui/app/(espace-pro)/espace-pro/(connected)/_components/FormulaireEditionOffreFields.tsx
@@ -14,10 +14,10 @@ import type { IAppellationsRomes } from "shared"
 import { NIVEAU_DIPLOME_LABEL, TRAINING_CONTRACT_TYPE } from "shared/constants/recruteur"
 import { JOB_START_TYPE } from "shared/models/job.model"
 import { AUTHTYPE } from "@/common/contants"
-import { debounce } from "@/common/utils/debounce"
 import { DropdownCombobox } from "@/components/espace_pro"
 import { useAuth } from "@/context/UserContext"
 import { apiGet } from "@/utils/api.utils"
+import { debounce } from "@/utils/debounce"
 import { ChampNombre } from "./ChampNombre"
 
 const ISO_DATE_FORMAT = "YYYY-MM-DD"
