+++ ui/app/(espace-pro)/espace-pro/(connected)/_components/OffresTabs.tsx
@@ -12,10 +12,10 @@ import { RECRUITER_STATUS } from "shared/constants/index"
 import Badge from "@/app/(espace-pro)/_components/Badge"
 import Table from "@/app/(espace-pro)/_components/Table"
 import { OffreProlongationModal } from "@/app/(espace-pro)/espace-pro/(connected)/_components/OffreProlongationModal"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { useToast } from "@/app/hooks/useToast"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
-import { sortReactTableDate } from "@/common/utils/date-utils"
 import { extendOffre } from "@/utils/api"
+import { sortReactTableDate } from "@/utils/date-utils"
 import ConfirmationSuppressionOffre from "./ConfirmationSuppressionOffre"
 import { OffresTabsMenu } from "./OffresTabsMenu"
