+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/gestion-des-administrateurs/gestionDesAdministrateurs.tsx
@@ -9,11 +9,11 @@ import { useRouter } from "next/navigation"
 import { Breadcrumb } from "@/app/_components/Breadcrumb"
 import LoadingEmptySpace from "@/app/(espace-pro)/_components/LoadingEmptySpace"
 import TableWithPagination from "@/app/(espace-pro)/_components/TableWithPagination"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
-import { sortReactTableString } from "@/common/utils/date-utils"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { ModalReadOnly } from "@/components/ModalReadOnly"
 import { ArrowRightLine } from "@/theme/components/icons"
 import { apiGet } from "@/utils/api.utils"
+import { sortReactTableString } from "@/utils/date-utils"
 import { PAGES } from "@/utils/routes.utils"
 import { AdminUserForm } from "./_components/AdminUserForm"
