+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/users/[userId]/cfa/[establishment_id]/AdminUserCfaEntreprisePage.tsx
@@ -7,7 +7,7 @@ import { Breadcrumb } from "@/app/_components/Breadcrumb"
 import LoadingEmptySpace from "@/app/(espace-pro)/_components/LoadingEmptySpace"
 import { ConfirmationSuppressionEntreprise } from "@/app/(espace-pro)/espace-pro/(connected)/_components/ConfirmationSuppressionEntreprise"
 import DetailEntreprise from "@/app/(espace-pro)/espace-pro/(connected)/_components/DetailEntreprise"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { getEntrepriseManagedByCfa, getFormulaire, getUser } from "@/utils/api"
 import { PAGES } from "@/utils/routes.utils"
 import { useSearchParamsRecord } from "@/utils/use-search-params-record"
