+++ ui/app/(espace-pro)/espace-pro/(connected)/opco/OpcoPage.tsx
@@ -8,14 +8,14 @@ import { useEffect, useState } from "react"
 import type { IUserRecruteurForAdminJSON } from "shared"
 import { Breadcrumb } from "@/app/_components/Breadcrumb"
 import TableWithPagination from "@/app/(espace-pro)/_components/TableWithPagination"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { useToast } from "@/app/hooks/useToast"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
-import { sortReactTableDate, sortReactTableString } from "@/common/utils/date-utils"
 import { ConfirmationDesactivationUtilisateur, LoadingEmptySpace } from "@/components/espace_pro"
 import ConfirmationActivationUtilisateur from "@/components/espace_pro/ConfirmationActivationUtilisateur"
 import { CustomTabs } from "@/components/espace_pro/CreationRecruteur/CustomTabs"
 import { webkitLineClamp } from "@/styles/webkit-line-clamp"
 import { getOpcoUsers } from "@/utils/api"
+import { sortReactTableDate, sortReactTableString } from "@/utils/date-utils"
 import { PAGES } from "@/utils/routes.utils"
 import { useSearchParamsRecord } from "@/utils/use-search-params-record"
 import { UserMenu } from "./_component/UserMenu"
