+++ ui/app/hooks/use-local-storage.ts
@@ -1,17 +1,23 @@
 import { useCallback, useState } from "react"
+import { localStorageGet, localStorageSet } from "@/utils/local-storage"
 
 export function useLocalStorage<T>(key: string, initialValue?: T) {
   const [storedValue, setStoredValue] = useState<T | null>(() => {
     if (typeof window === "undefined") return initialValue ?? null
-    const item = localStorage.getItem(key)
-    return item ? JSON.parse(item) : (initialValue ?? null)
+    const item = localStorageGet(key)
+    if (!item) return initialValue ?? null
+    try {
+      return JSON.parse(item)
+    } catch {
+      return initialValue ?? null
+    }
   })
 
   // Function to update localStorage
   const setLocalStorage = useCallback(
     (value: T) => {
       if (typeof window === "undefined") return
-      localStorage.setItem(key, JSON.stringify(value))
+      localStorageSet(key, value)
       setStoredValue(value)
     },
     [key]
