+++ ui/components/ItemDetail/PartnerJobComponents/PartnerJobExternalApply.tsx
@@ -5,7 +5,7 @@ import type { ILbaItemPartnerJobJson } from "shared"
 import { JOBPARTNERS_LABEL } from "shared/models/jobs-partners.model"
 import ModalCloseButton from "@/app/_components/ModalCloseButton"
 import { ModalTitle } from "@/app/_components/Title/ModalTitle"
-import { useDisclosure } from "@/common/hooks/use-disclosure"
+import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { notifyJobPostulerV3 } from "@/utils/api"
 import { MATOMO_EVENTS, pushMatomoEvent } from "@/utils/matomo-utils"
 import { SendPlausibleEvent } from "@/utils/plausible"
