+++ ui/components/espace_pro/ConfirmationActivationUtilisateur.tsx
@@ -2,7 +2,7 @@ import { fr } from "@codegouvfr/react-dsfr"
 import Button from "@codegouvfr/react-dsfr/Button"
 import { Box, Typography } from "@mui/material"
 
-import { useUserPermissionsActions } from "@/common/hooks/use-user-permissions-actions"
+import { useUserPermissionsActions } from "@/app/hooks/use-user-permissions-actions"
 import { ModalReadOnly } from "@/components/ModalReadOnly"
 
 const ConfirmationActivationUtilisateur = (props: {
